Faet, style, repectoring: 거북목 스트레칭 모드 완료, 주석제거 및 코드스타일 정검(필요없는 코드 삭제), 자세에 대한 알고리즘은 추가적으로 조정이 필요
