+++ src/app/GlobalStyles.js
@@ -4,7 +4,6 @@ import reset from 'styled-reset';
 const GlobalStyles = createGlobalStyle`
   ${reset}
 
-  /* 모든 엘리먼트의 box-sizing 값을 border-box로 설정 */
   * {
     box-sizing: inherit;
   }
@@ -16,12 +15,10 @@ const GlobalStyles = createGlobalStyle`
      'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue', sans-serif;
     -webkit-font-smoothing: antialiased;
     -moz-osx-font-smoothing: grayscale;
-    /* 엘리먼트의 box-sizing 값을 border-box로 설정 */
     box-sizing: border-box;
     min-height: 100%;
   }
 
-  /* 링크에 색상 및 밑줄 없애기 */
   a {
     color: inherit;
     text-decoration: none;
