--- src/components/ProgressBar.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const ProgressBar = () => {
-  return <div>ProgressBar</div>;
-};
-
-export default ProgressBar;
