+++ src/components/TextCard.js
@@ -9,8 +9,8 @@ const TextCard = () => {
         위의 스트레칭 자세를 참고하여 스트레칭을 진행해 주세요.
       </div>
       <div className='text-area-description'>
-        두 팔을 들어 120'c 각도를 유지한 채로 날개뼈를 모으고 고개를 들면서 팔
-        안쪽이 당기는 느낌을 받습니다. 10초 유지합니다.
+        두 팔을 들어 120'c 각도를 유지한 채로 날개뼈를 모으고 고개를 들면서 팔을
+        안쪽으로 당기는 느낌을 받으며 당겨줍니다. 그 자세로 10초 유지합니다.
       </div>
     </CardWrap>
   );
