+++ src/util/helpers/utils.js
@@ -1,0 +1,126 @@
+import * as poseNet from '@tensorflow-models/posenet';
+import * as tf from '@tensorflow/tfjs';
+
+const color = '#75fe4b';
+const boundingBoxColor = 'red';
+const lineWidth = 5;
+
+export const tryResNetButtonName = 'tryResNetButton';
+export const tryResNetButtonText = '[New] Try ResNet50';
+const tryResNetButtonTextCss = 'width:100%;text-decoration:underline;';
+const tryResNetButtonBackgroundCss = 'background:#e61d5f;';
+
+function toTuple({ y, x }) {
+  return [y, x];
+}
+
+export function drawPoint(ctx, y, x, r, color) {
+  ctx.beginPath();
+  ctx.arc(x, y, r, 0, 2 * Math.PI);
+  ctx.fillStyle = color;
+  ctx.fill();
+}
+
+export function drawSegment([ay, ax], [by, bx], color, scale, ctx) {
+  ctx.beginPath();
+  ctx.moveTo(ax * scale, ay * scale);
+  ctx.lineTo(bx * scale, by * scale);
+  ctx.lineWidth = lineWidth;
+  ctx.strokeStyle = color;
+  ctx.stroke();
+}
+
+export function drawSkeleton(keypoints, minConfidence, ctx, scale = 1) {
+  const adjacentKeyPoints = poseNet.getAdjacentKeyPoints(
+    keypoints,
+    minConfidence,
+  );
+
+  adjacentKeyPoints.forEach((keypoints) => {
+    drawSegment(
+      toTuple(keypoints[0].position),
+      toTuple(keypoints[1].position),
+      color,
+      scale,
+      ctx,
+    );
+  });
+}
+
+export function drawKeyPoints(keypoints, minConfidence, ctx, scale = 1) {
+  for (let i = 0; i < keypoints.length; i++) {
+    const keypoint = keypoints[i];
+
+    if (keypoint.score < minConfidence) {
+      continue;
+    }
+
+    const { y, x } = keypoint.position;
+    drawPoint(ctx, y * scale, x * scale, 3, color);
+  }
+}
+
+export function drawBoundingBox(keypoints, ctx) {
+  const boundingBox = poseNet.getBoundingBox(keypoints);
+
+  ctx.rect(
+    boundingBox.minX,
+    boundingBox.minY,
+    boundingBox.maxX - boundingBox.minX,
+    boundingBox.maxY - boundingBox.minY,
+  );
+
+  ctx.strokeStyle = boundingBoxColor;
+  ctx.stroke();
+}
+
+export async function renderToCanvas(a, ctx) {
+  const [height, width] = a.shape;
+  const imageData = new ImageData(width, height);
+
+  const data = await a.data();
+
+  for (let i = 0; i < height * width; ++i) {
+    const j = i * 4;
+    const k = i * 3;
+
+    imageData.data[j + 0] = data[k + 0];
+    imageData.data[j + 1] = data[k + 1];
+    imageData.data[j + 2] = data[k + 2];
+    imageData.data[j + 3] = 255;
+  }
+
+  ctx.putImageData(imageData, 0, 0);
+}
+
+export function renderImageToCanvas(image, size, canvas) {
+  canvas.width = size[0];
+  canvas.height = size[1];
+  const ctx = canvas.getContext('2d');
+
+  ctx.drawImage(image, 0, 0);
+}
+
+export function drawHeatMapValues(heatMapValues, outputStride, canvas) {
+  const ctx = canvas.getContext('2d');
+  const radius = 5;
+  const scaledValues = heatMapValues.mul(tf.scalar(outputStride, 'int32'));
+
+  drawPoints(ctx, scaledValues, radius, color);
+}
+
+function drawPoints(ctx, points, radius, color) {
+  const data = points.buffer().values;
+
+  for (let i = 0; i < data.length; i += 2) {
+    const pointY = data[i];
+    const pointX = data[i + 1];
+
+    if (pointX !== 0 && pointY !== 0) {
+      ctx.beginPath();
+      ctx.arc(pointX, pointY, radius, 0, 2 * Math.PI);
+      ctx.fillStyle = color;
+      ctx.fill();
+    }
+  }
+}
